+++ src/pages/Mypage.jsx
@@ -1,7 +1,7 @@
 import {useEffect, useState} from "react";
 import { useNavigate } from 'react-router-dom';
 import axios from "axios";
-import "./Mypage.css";
+import "../styles/Mypage.css";
 
 const Mypage = () => {
     const [userInfo, setUserInfo] = useState(null);
